refactor: extract REST API prompt builders into sessionApiPrompts.js
Split session prompt generation to separate API-related builders into a
dedicated module, reducing sessionPrompts.js size and improving cohesion.
Also adds dedicated POST /api/sessions/:id/schedule endpoint support and
minor UI fixes in ConversationTab and ScheduleSessionModal.
